Refactoring application
