+++ app/src/main/java/es/um/comov/p2/Utils.java
@@ -14,26 +14,27 @@
  * limitations under the License.
  */
 
-package com.google.android.gms.location.sample.locationupdatesforegroundservice;
+package es.um.comov.p2;
 
 
 import android.content.Context;
 import android.location.Location;
 import android.preference.PreferenceManager;
 
+import es.um.comov.p2.R;
+
 import java.text.DateFormat;
 import java.util.Date;
+public class Utils {
 
-class Utils {
-
-    static final String KEY_REQUESTING_LOCATION_UPDATES = "requesting_location_updates";
-
+    public static final String KEY_REQUESTING_LOCATION_UPDATES = "requesting_location_updates";
+    
     /**
      * Returns true if requesting location updates, otherwise returns false.
      *
      * @param context The {@link Context}.
      */
-    static boolean requestingLocationUpdates(Context context) {
+    public static boolean requestingLocationUpdates(Context context) {
         return PreferenceManager.getDefaultSharedPreferences(context)
                 .getBoolean(KEY_REQUESTING_LOCATION_UPDATES, false);
     }
@@ -42,7 +43,7 @@ class Utils {
      * Stores the location updates state in SharedPreferences.
      * @param requestingLocationUpdates The location updates state.
      */
-    static void setRequestingLocationUpdates(Context context, boolean requestingLocationUpdates) {
+    public static void setRequestingLocationUpdates(Context context, boolean requestingLocationUpdates) {
         PreferenceManager.getDefaultSharedPreferences(context)
                 .edit()
                 .putBoolean(KEY_REQUESTING_LOCATION_UPDATES, requestingLocationUpdates)
@@ -53,12 +54,12 @@ class Utils {
      * Returns the {@code location} object as a human readable string.
      * @param location  The {@link Location}.
      */
-    static String getLocationText(Location location) {
+    public static String getLocationText(Location location) {
         return location == null ? "Unknown location" :
                 "(" + location.getLatitude() + ", " + location.getLongitude() + ")";
     }
 
-    static String getLocationTitle(Context context) {
+    public static String getLocationTitle(Context context) {
         return context.getString(R.string.location_updated,
                 DateFormat.getDateTimeInstance().format(new Date()));
     }
