+++ app/src/main/res/layout/activity_coverage_map.xml
@@ -6,4 +6,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CoverageMapActivity" />
+    tools:context="es.um.comov.p2.CoverageMapActivity" />
