+++ app/src/main/res/layout/activity_main.xml
@@ -23,7 +23,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.google.android.gms.location.sample.locationupdatesforegroundservice.MainActivity">
+    tools:context="com.google.android.gms.location.sample.locationupdatesforegroundservice.es.um.comov.p2.MainActivity">
 
     <Button
         android:id="@+id/request_location_updates_button"
